python/hello-cuda/pyproject.toml
[tool.poetry]
name = "hello-cuda"
version = "0.1.0"
description = ""
authors = ["Akagi201 <akagi201@gmail.com>"]
readme = "README.md"
packages = [{include = "hello_cuda", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
cupy-cuda12x = "^13.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
